requirements.txt
Flask
SQLAlchemy
psycopg2-binary
Alembic
python-dotenv
pyjwt
Flask-JWT-Extended